+++ freekick/api/__init__.py
@@ -6,6 +6,7 @@ from .gameweek import gameweek_ns
 from .healthcheck import HealthCheckApi
 from .match import match_ns
 from .season import season_ns
+from .settings import setting_ns
 
 freekick_api = Api(
     title="Freekick",
@@ -18,4 +19,5 @@ freekick_api = Api(
 freekick_api.add_namespace(match_ns)
 freekick_api.add_namespace(gameweek_ns)
 freekick_api.add_namespace(season_ns)
+freekick_api.add_namespace(setting_ns)
 freekick_api.add_resource(HealthCheckApi, "/health")