+++ freekick/learners/__init__.py
@@ -3,9 +3,11 @@ from freekick.learners.learner_utils import (
     AllEstimator,
     DEFAULT_ESTIMATOR,
 )
+from .classification import FreekickDecisionTreeClassifier
 
 __all__ = [
     "serial_models",
     "AllEstimator",
-    "DEFAULT_ESTIMATOR",
+    "DEFAULT_ESTIMATOR",  # TODO: Remove all use of DEFAULT_ESTIMATOR, use specific estimators
+    "FreekickDecisionTreeClassifier"
 ]