+++ freekick/service/__init__.py
@@ -1,11 +1,16 @@
 from .match_day_predictor import predict_match_day
 from .match_predictor import predict_match
 from .season_service import get_current_season_teams
-from .util import MatchDTO
+from .setting_service import get_setting, update_setting
+from .util import MatchDTO, SettingDTO, SeasonDTO
 
 __all__ = [
     "predict_match_day",
     "predict_match",
     "MatchDTO",
+    "SettingDTO",
+    "SeasonDTO",
     "get_current_season_teams",
+    "get_setting",
+    "update_setting",
 ]